@@ -16,39 +16,33 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
-        "symfony/deprecation-contracts": "^2.1|^3",
+        "php": ">=8.2",
         "symfony/polyfill-mbstring": "~1.0",
-        "symfony/polyfill-php73": "^1.9",
-        "symfony/polyfill-php80": "^1.16",
-        "symfony/service-contracts": "^1.1|^2|^3",
-        "symfony/string": "^5.1|^6.0"
+        "symfony/service-contracts": "^2.5|^3",
+        "symfony/string": "^6.4|^7.0"
     },
     "require-dev": {
-        "symfony/config": "^4.4|^5.0|^6.0",
-        "symfony/event-dispatcher": "^4.4|^5.0|^6.0",
-        "symfony/dependency-injection": "^4.4|^5.0|^6.0",
-        "symfony/lock": "^4.4|^5.0|^6.0",
-        "symfony/process": "^4.4|^5.0|^6.0",
-        "symfony/var-dumper": "^4.4|^5.0|^6.0",
-        "psr/log": "^1|^2"
+        "symfony/config": "^6.4|^7.0",
+        "symfony/event-dispatcher": "^6.4|^7.0",
+        "symfony/http-foundation": "^6.4|^7.0",
+        "symfony/http-kernel": "^6.4|^7.0",
+        "symfony/dependency-injection": "^6.4|^7.0",
+        "symfony/lock": "^6.4|^7.0",
+        "symfony/messenger": "^6.4|^7.0",
+        "symfony/process": "^6.4|^7.0",
+        "symfony/stopwatch": "^6.4|^7.0",
+        "symfony/var-dumper": "^6.4|^7.0",
+        "psr/log": "^1|^2|^3"
     },
     "provide": {
-        "psr/log-implementation": "1.0|2.0"
-    },
-    "suggest": {
-        "symfony/event-dispatcher": "",
-        "symfony/lock": "",
-        "symfony/process": "",
-        "psr/log": "For using the console logger"
+        "psr/log-implementation": "1.0|2.0|3.0"
     },
     "conflict": {
-        "psr/log": ">=3",
-        "symfony/dependency-injection": "<4.4",
-        "symfony/dotenv": "<5.1",
-        "symfony/event-dispatcher": "<4.4",
-        "symfony/lock": "<4.4",
-        "symfony/process": "<4.4"
+        "symfony/dependency-injection": "<6.4",
+        "symfony/dotenv": "<6.4",
+        "symfony/event-dispatcher": "<6.4",
+        "symfony/lock": "<6.4",
+        "symfony/process": "<6.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Console\\": "" },
