@@ -16,13 +16,12 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.2",
         "symfony/polyfill-ctype": "~1.8",
-        "symfony/polyfill-mbstring": "~1.8",
-        "symfony/polyfill-php80": "^1.16"
+        "symfony/polyfill-mbstring": "~1.8"
     },
     "require-dev": {
-        "symfony/process": "^5.4|^6.4"
+        "symfony/process": "^6.4|^7.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Filesystem\\": "" },
