@@ -16,9 +16,10 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
-        "symfony/deprecation-contracts": "^2.1|^3",
-        "symfony/polyfill-php80": "^1.16"
+        "php": ">=8.2"
+    },
+    "require-dev": {
+        "symfony/filesystem": "^6.4|^7.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Finder\\": "" },
