@@ -16,7 +16,7 @@
         }
     ],
     "require": {
-        "php": ">=7.1"
+        "php": ">=7.2"
     },
     "autoload": {
         "psr-4": { "Symfony\\Polyfill\\Intl\\Normalizer\\": "" },
