@@ -43,7 +43,32 @@ require_once ABSPATH . 'wp-admin/admin-header.php';
 
 		<div class="about__section changelog has-subtle-background-color">
 			<div class="column">
-				<h2><?php _e( 'Maintenance and Security Release' ); ?></h2>
+				<h2><?php _e( 'Maintenance and Security Releases' ); ?></h2>
+				<p>
+					<?php
+					printf(
+						/* translators: 1: WordPress version number, 2: Plural number of bugs. */
+						_n(
+							'<strong>Version %1$s</strong> addressed %2$s bug.',
+							'<strong>Version %1$s</strong> addressed %2$s bugs.',
+							26
+						),
+						'6.6.2',
+						'26'
+					);
+					?>
+					<?php
+					printf(
+						/* translators: %s: HelpHub URL. */
+						__( 'For more information, see <a href="%s">the release notes</a>.' ),
+						sprintf(
+							/* translators: %s: WordPress version. */
+							esc_url( __( 'https://wordpress.org/support/wordpress-version/version-%s/' ) ),
+							sanitize_title( '6.6.2' )
+						)
+					);
+					?>
+				</p>
 				<p>
 					<?php
 					printf(
