@@ -2,8 +2,12 @@
 /**
  * Pop-up displayed on plugin settings page.
  *
- * @var string $author_image_url    Avatar of plugin author.
- * @var bool   $token_active_status Status of PRO version.
+ * @var string $author_image_url         Avatar of plugin author.
+ * @var string $form_input_name          Name of hidden field with form ID.
+ * @var string $form_sidebar_input_value ID of settings form.
+ * @var string $nonce_input_name         Name of hidden field with WordPress Nonce value.
+ * @var string $nonce_input_value        WordPress Nonce value.
+ * @var bool   $token_active_status      Status of PRO version.
  *
  * @package Converter for Media
  */
@@ -14,11 +18,10 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 ?>
 <?php if ( ! $token_active_status ) : ?>
-	<div class="webpcPopup" hidden data-popup="regeneration">
+	<div class="webpcPopup" data-popup="regeneration" hidden>
 		<div class="webpcPopup__wrapper">
-			<div class="webpcPopup__inner" data-popup-content>
-				<button type="button" class="webpcPopup__close dashicons dashicons-no"
-					data-popup-close></button>
+			<div class="webpcPopup__inner" data-popup-page="1">
+				<button type="button" class="webpcPopup__close dashicons dashicons-no" data-popup-close></button>
 				<div class="webpcPopup__sidebar">
 					<div class="webpcPopup__sidebarInner">
 						<img src="<?php echo esc_attr( $author_image_url ); ?>" alt="" class="webpcPopup__sidebarImage">
@@ -27,26 +30,105 @@ if ( ! defined( 'ABSPATH' ) ) {
 				</div>
 				<div class="webpcPopup__content">
 					<div class="webpcPopup__contentText">
-						<?php
-						echo esc_html(
-							sprintf(
-							/* translators: %1$s: author name, %2$s: format name, %3$s: percent value, %4$s: format name */
-								__( 'Hi - I am %1$s, the author of this plugin. Did you know that by converting your images to the %2$s format as well, you can reduce the weight of your images by an additional about %3$s compared to using only the %4$s format?', 'webp-converter-for-media' ),
-								'Mateusz',
-								'AVIF',
-								'50%',
-								'WebP'
-							)
-						);
-						?>
+						<p>
+							<?php
+							echo esc_html(
+								sprintf(
+								/* translators: %s: author name */
+									__( 'Hi - I am %s, the author of this plugin.', 'webp-converter-for-media' ),
+									'Mateusz'
+								)
+							);
+							?>
+							<strong>
+								<?php
+								echo esc_html(
+									sprintf(
+									/* translators: %s: format name */
+										__( 'Please tell me, would you like to know more about the %s format?', 'webp-converter-for-media' ),
+										'AVIF'
+									)
+								);
+								?>
+							</strong>
+						</p>
+						<p>
+							<?php echo esc_html__( 'The AVIF format is the successor to the WebP format. Images converted to the AVIF format weigh about 50% less than images converted only to the WebP format, while maintaining better image quality.', 'webp-converter-for-media' ); ?>
+						</p>
 					</div>
-					<div class="webpcPopup__contentButton">
-						<a href="https://url.mattplugins.com/converter-regeneration-popup-avif" target="_blank"
-							class="webpcButton webpcButton--blue webpcButton--bg">
-							<?php echo esc_html( __( 'Explore the opportunities for yourself', 'webp-converter-for-media' ) ); ?>
-						</a>
+					<div class="webpcPopup__contentButtons">
+						<div class="webpcPopup__contentButton">
+							<a href="https://url.mattplugins.com/converter-regeneration-popup-step-avif" target="_blank"
+								class="webpcButton webpcButton--blue webpcButton--bg"
+								data-popup-button-page="2">
+								<?php echo esc_html__( 'Find out more', 'webp-converter-for-media' ); ?>
+							</a>
+						</div>
+						<div class="webpcPopup__contentButton">
+							<button type="button"
+								class="webpcButton webpcButton--gray webpcButton--bg"
+								data-popup-button-page="2">
+								<?php echo esc_html__( 'Skip for now', 'webp-converter-for-media' ); ?>
+							</button>
+						</div>
 					</div>
 				</div>
+			</div>
+			<div class="webpcPopup__inner" data-popup-page="2" hidden>
+				<button type="button" class="webpcPopup__close dashicons dashicons-no" data-popup-close></button>
+				<form method="post" action="" class="webpcPopup__content">
+					<input type="hidden" name="<?php echo esc_attr( $form_input_name ); ?>"
+						value="<?php echo esc_attr( $form_sidebar_input_value ); ?>">
+					<input type="hidden" name="<?php echo esc_attr( $nonce_input_name ); ?>"
+						value="<?php echo esc_attr( $nonce_input_value ); ?>">
+					<div class="webpcPopup__contentText">
+						<p>
+							<strong>
+								<?php
+								echo esc_html(
+									sprintf(
+									/* translators: %s: format name */
+										__( 'Provide a valid access token to continue the process of optimizing images to the %s format.', 'webp-converter-for-media' ),
+										'AVIF'
+									)
+								);
+								?>
+							</strong>
+						</p>
+						<p>
+							<?php echo esc_html__( 'Converting images to WebP and AVIF simultaneously guarantees the lowest weight of your images and compatibility with all browsers. By using the AVIF format you will reduce the weight of your images even more compared to WebP.', 'webp-converter-for-media' ); ?>
+						</p>
+						<p data-plans>
+							<?php
+							echo wp_kses_post(
+								sprintf(
+								/* translators: %1$s: open anchor tag, %2$s: close anchor tag */
+									__( 'You can get your access token %1$shere%2$s.', 'webp-converter-for-media' ),
+									'<a href="https://url.mattplugins.com/converter-regeneration-popup-step-token?webp=0&avif=0" target="_blank" data-plans-button>',
+									'</a>'
+								)
+							);
+							?>
+						</p>
+					</div>
+					<div class="webpcPopup__input webpcInput">
+						<input type="text" name="access_token" id="access_token" class="webpcInput__field" required>
+					</div>
+					<div class="webpcPopup__contentButtons">
+						<div class="webpcPopup__contentButton">
+							<button type="submit" class="webpcButton webpcButton--blue webpcButton--bg">
+								<?php echo esc_html__( 'Activate Token', 'webp-converter-for-media' ); ?>
+							</button>
+						</div>
+						<div class="webpcPopup__contentButton">
+							<button type="button"
+								class="webpcButton webpcButton--gray webpcButton--bg"
+								data-popup-close>
+								<?php echo esc_html__( 'Not now', 'webp-converter-for-media' ); ?>
+							</button>
+						</div>
+					</div>
+				</form>
 			</div>
 		</div>
 	</div>
