@@ -3,7 +3,7 @@
 /**
  * Plugin Name: Converter for Media
  * Description: Speed up your website by using our WebP & AVIF Converter (formerly WebP Converter for Media). Serve WebP and AVIF images instead of standard formats JPEG, PNG and GIF now!
- * Version: 6.0.0
+ * Version: 6.1.0
  * Author: matt plugins - Optimize images by convert WebP & AVIF
  * Author URI: https://url.mattplugins.com/converter-plugin-author-link
  * Text Domain: webp-converter-for-media
@@ -17,5 +17,5 @@ if ( ! defined( 'ABSPATH' ) ) {
 require_once __DIR__ . '/vendor/autoload.php';
 
 new WebpConverter\WebpConverter(
-	new WebpConverter\PluginInfo( __FILE__, '6.0.0' )
+	new WebpConverter\PluginInfo( __FILE__, '6.1.0' )
 );
