@@ -168,24 +168,10 @@ class WPSEO_Option_Social extends WPSEO_Option {
 						if ( $twitter_id ) {
 							$clean[ $key ] = $twitter_id;
 						}
-						else {
-							if ( isset( $old[ $key ] ) && $old[ $key ] !== '' ) {
+						elseif ( isset( $old[ $key ] ) && $old[ $key ] !== '' ) {
 								$twitter_id = sanitize_text_field( ltrim( $old[ $key ], '@' ) );
-								if ( preg_match( '`^[A-Za-z0-9_]{1,25}$`', $twitter_id ) ) {
-									$clean[ $key ] = $twitter_id;
-								}
-							}
-							if ( function_exists( 'add_settings_error' ) ) {
-								add_settings_error(
-									$this->group_name, // Slug title of the setting.
-									$key, // Suffix-ID for the error message box.
-									sprintf(
-										/* translators: %s expands to a twitter user name. */
-										__( '%s does not seem to be a valid Twitter Username. Please correct.', 'wordpress-seo' ),
-										'<strong>' . esc_html( sanitize_text_field( $dirty[ $key ] ) ) . '</strong>'
-									), // The error message.
-									'error' // Message type.
-								);
+							if ( preg_match( '`^[A-Za-z0-9_]{1,25}$`', $twitter_id ) ) {
+								$clean[ $key ] = $twitter_id;
 							}
 						}
 						unset( $twitter_id );
