@@ -34,7 +34,14 @@ class Primary_Category_Conditional implements Conditional {
 	 *          post edit, new post admin page or additional admin pages, allowed by filter.
 	 */
 	public function is_met() {
+
 		if ( ! \is_admin() ) {
+			return true;
+		}
+
+		$current_page = $this->current_page->get_current_admin_page();
+		// phpcs:ignore WordPress.Security.NonceVerification.Missing -- Reason: We are not processing form information.
+		if ( $current_page === 'admin-ajax.php' && isset( $_POST['action'] ) && $_POST['action'] === 'wp-link-ajax' ) {
 			return true;
 		}
 
@@ -44,6 +51,6 @@ class Primary_Category_Conditional implements Conditional {
 		 * @param array $admin_pages List of additional admin pages.
 		 */
 		$additional_pages = \apply_filters( 'wpseo_primary_category_admin_pages', [] );
-		return \in_array( $this->current_page->get_current_admin_page(), \array_merge( [ 'edit.php', 'post.php', 'post-new.php' ], $additional_pages ), true );
+		return \in_array( $current_page, \array_merge( [ 'edit.php', 'post.php', 'post-new.php' ], $additional_pages ), true );
 	}
 }
