@@ -90,7 +90,11 @@ class Website extends Abstract_Schema_Piece {
 				'@type'       => 'EntryPoint',
 				'urlTemplate' => $search_url,
 			],
-			'query-input' => 'required name=search_term_string',
+			'query-input' => [
+				'@type'          => 'PropertyValueSpecification',
+				'valueRequired'  => true,
+				'valueName'      => 'search_term_string',
+			],
 		];
 
 		return $data;
