@@ -6,22 +6,29 @@ use Easy_Digital_Downloads;
 use SeriouslySimplePodcasting\Integrations\Yoast\Schema\PodcastEpisode;
 use TEC\Events\Integrations\Plugins\WordPress_SEO\Events_Schema;
 use WP_Recipe_Maker;
+use WPSEO_Addon_Manager;
 use WPSEO_Admin_Asset_Manager;
-use WPSEO_Plugin_Availability;
 use WPSEO_Shortlinker;
 use Yoast\WP\SEO\Conditionals\Admin_Conditional;
 use Yoast\WP\SEO\Conditionals\Jetpack_Conditional;
 use Yoast\WP\SEO\Conditionals\Third_Party\Elementor_Activated_Conditional;
 use Yoast\WP\SEO\Conditionals\Third_Party\Jetpack_Boost_Active_Conditional;
 use Yoast\WP\SEO\Conditionals\Third_Party\Jetpack_Boost_Not_Premium_Conditional;
-use Yoast\WP\SEO\Conditionals\WooCommerce_Conditional;
 use Yoast\WP\SEO\Helpers\Options_Helper;
+use Yoast\WP\SEO\Helpers\Woocommerce_Helper;
 use Yoast\WP\SEO\Integrations\Integration_Interface;
 
 /**
  * Integrations_Page class
  */
 class Integrations_Page implements Integration_Interface {
+
+	/**
+	 * The Woocommerce helper.
+	 *
+	 * @var Woocommerce_Helper
+	 */
+	private $woocommerce_helper;
 
 	/**
 	 * The admin asset manager.
@@ -49,13 +56,16 @@ class Integrations_Page implements Integration_Interface {
 	 *
 	 * @param WPSEO_Admin_Asset_Manager $admin_asset_manager The admin asset manager.
 	 * @param Options_Helper            $options_helper      The options helper.
+	 * @param Woocommerce_Helper        $woocommerce_helper  The WooCommerce helper.
 	 */
 	public function __construct(
 		WPSEO_Admin_Asset_Manager $admin_asset_manager,
-		Options_Helper $options_helper
+		Options_Helper $options_helper,
+		Woocommerce_Helper $woocommerce_helper
 	) {
 		$this->admin_asset_manager = $admin_asset_manager;
 		$this->options_helper      = $options_helper;
+		$this->woocommerce_helper  = $woocommerce_helper;
 	}
 
 	/**
@@ -107,7 +117,6 @@ class Integrations_Page implements Integration_Interface {
 
 		$elementor_conditional                 = new Elementor_Activated_Conditional();
 		$jetpack_conditional                   = new Jetpack_Conditional();
-		$woocommerce_conditional               = new WooCommerce_Conditional();
 		$jetpack_boost_active_conditional      = new Jetpack_Boost_Active_Conditional();
 		$jetpack_boost_not_premium_conditional = new Jetpack_Boost_Not_Premium_Conditional();
 
@@ -117,26 +126,26 @@ class Integrations_Page implements Integration_Interface {
 		$algolia_file         = 'wp-search-with-algolia/algolia.php';
 		$old_algolia_file     = 'search-by-algolia-instant-relevant-results/algolia.php';
 
-		$host = \YoastSEO()->helpers->url->get_url_host( \get_site_url() );
+		$host                      = \YoastSEO()->helpers->url->get_url_host( \get_site_url() );
+		$addon_manager             = new WPSEO_Addon_Manager();
+		$woocommerce_seo_installed = $addon_manager->is_installed( WPSEO_Addon_Manager::WOOCOMMERCE_SLUG );
 
-		$wpseo_plugin_availability_checker = new WPSEO_Plugin_Availability();
-		$woocommerce_seo_installed         = \file_exists( \WP_PLUGIN_DIR . '/' . $woocommerce_seo_file );
-		$woocommerce_seo_active            = $wpseo_plugin_availability_checker->is_active( $woocommerce_seo_file );
-		$woocommerce_active                = $woocommerce_conditional->is_met();
-		$acf_seo_installed                 = \file_exists( \WP_PLUGIN_DIR . '/' . $acf_seo_file );
-		$acf_seo_github_installed          = \file_exists( \WP_PLUGIN_DIR . '/' . $acf_seo_file_github );
-		$acf_seo_active                    = $wpseo_plugin_availability_checker->is_active( $acf_seo_file );
-		$acf_seo_github_active             = $wpseo_plugin_availability_checker->is_active( $acf_seo_file_github );
-		$acf_active                        = \class_exists( 'acf' );
-		$algolia_active                    = $wpseo_plugin_availability_checker->is_active( $algolia_file );
-		$edd_active                        = \class_exists( Easy_Digital_Downloads::class );
-		$jetpack_boost_active              = $jetpack_boost_active_conditional->is_met();
-		$jetpack_boost_premium             = ( ! $jetpack_boost_not_premium_conditional->is_met() );
-		$old_algolia_active                = $wpseo_plugin_availability_checker->is_active( $old_algolia_file );
-		$tec_active                        = \class_exists( Events_Schema::class );
-		$ssp_active                        = \class_exists( PodcastEpisode::class );
-		$wp_recipe_maker_active            = \class_exists( WP_Recipe_Maker::class );
-		$mastodon_active                   = $this->is_mastodon_active();
+		$woocommerce_seo_active   = \is_plugin_active( $woocommerce_seo_file );
+		$woocommerce_active       = $this->woocommerce_helper->is_active();
+		$acf_seo_installed        = \file_exists( \WP_PLUGIN_DIR . '/' . $acf_seo_file );
+		$acf_seo_github_installed = \file_exists( \WP_PLUGIN_DIR . '/' . $acf_seo_file_github );
+		$acf_seo_active           = \is_plugin_active( $acf_seo_file );
+		$acf_seo_github_active    = \is_plugin_active( $acf_seo_file_github );
+		$acf_active               = \class_exists( 'acf' );
+		$algolia_active           = \is_plugin_active( $algolia_file );
+		$edd_active               = \class_exists( Easy_Digital_Downloads::class );
+		$jetpack_boost_active     = $jetpack_boost_active_conditional->is_met();
+		$jetpack_boost_premium    = ( ! $jetpack_boost_not_premium_conditional->is_met() );
+		$old_algolia_active       = \is_plugin_active( $old_algolia_file );
+		$tec_active               = \class_exists( Events_Schema::class );
+		$ssp_active               = \class_exists( PodcastEpisode::class );
+		$wp_recipe_maker_active   = \class_exists( WP_Recipe_Maker::class );
+		$mastodon_active          = $this->is_mastodon_active();
 
 		$woocommerce_seo_activate_url = \wp_nonce_url(
 			\self_admin_url( 'plugins.php?action=activate&plugin=' . $woocommerce_seo_file ),
