@@ -1,30 +1,30 @@
 <div class="wrap">
 
-    <div id="swcfpc_main_content">
+	<div id="swcfpc_main_content">
 
-        <h1><?php _e('Super Page Cache - Cached HTML pages', 'wp-cloudflare-page-cache'); ?></h1>
+		<h1><?php _e( 'Super Page Cache - Cached HTML pages', 'wp-cloudflare-page-cache' ); ?></h1>
 
 
-        <?php if( count($cached_html_pages_list) > 0 ): ?>
+		<?php if ( count( $cached_html_pages_list ) > 0 ) : ?>
 
-            <p><strong><?php printf( __('There are %d cached pages in list:', 'wp-cloudflare-page-cache'), count($cached_html_pages_list) ); ?></strong></p>
+			<p><strong><?php printf( __( 'There are %d cached pages in list:', 'wp-cloudflare-page-cache' ), count( $cached_html_pages_list ) ); ?></strong></p>
 
-            <ul>
+			<ul>
 
-                <?php foreach($cached_html_pages_list as $url): ?>
+				<?php foreach ( $cached_html_pages_list as $url ) : ?>
 
-                    <li><?php echo $url; ?></li>
+					<li><?php echo $url; ?></li>
 
-                <?php endforeach; ?>
+				<?php endforeach; ?>
 
-            </ul>
+			</ul>
 
-        <?php else: ?>
+		<?php else : ?>
 
-            <p><?php echo __( 'There are no cached HTML pages right now.', 'wp-cloudflare-page-cache' ); ?></p>
+			<p><?php echo __( 'There are no cached HTML pages right now.', 'wp-cloudflare-page-cache' ); ?></p>
 
-        <?php endif; ?>
+		<?php endif; ?>
 
-    </div>
+	</div>
 
 </div>
